+++ src/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Home from './Home/Home';
+import App  from './App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
-    <Home />,
+    <App />,
   document.getElementById('root')
 );
